Move `internal/client` to `pkg/client`
Make the client package public to be re-used from other projects.
